requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
early-stopping-pytorch==1.0.10
filelock==3.16.1
fsspec==2024.10.0
idna==3.10
Jinja2==3.1.4
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.1.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-python==4.10.0.84
pillow==11.0.0
redis==5.2.0
requests==2.32.3
spotipy==2.24.0
sympy==1.13.1
torch==2.5.1
torchvision==0.20.1
triton==3.1.0
typing_extensions==4.12.2
urllib3==2.2.3
